Fix a race condition in starting up tunnel sessions
This commit fixes a race condition where AsyncSSH could hang when
waiting for a connection to close if a request to open a new TCP or UNIX
domain channel came in after the connection was closed but before the
wait_closed was called on it. Such channel requests weren't returning
an error after connection close, leaving a task waiting for a reply to
the open request that would never come.