Miscellaneous cleanup for issues found during unit testing
This commit adds an exception handler in create_connection() which
cleans up the SSHAgentClient object it creates if an error occurs
later while opening the connection, avoiding a resource warning
about an unclosed socket.

This commit also removes unnecessary encoding calls in SSHConnection
and SSHChannel, and an unnecessary conditional in the get_key_options()
and check_key_permissions() methods.

It also updates the documentation of check_certificate_permissions()
to indicate that agent-forwarding permissions are now enforced.