Move algorithm validity checks out of SSHConnection constructor
This commit moves the validation of algorithm names provided by the
caller from the SSHConnection constructor to the create_connection()
and create_server() functions to avoid a bug in asyncio where a socket
object is leaked if a protocol factory raises an exception.

A benefit of this change is that invalid algorithms are now caught at
the time create_server() is called, rather than only being detected when
a client attempts to connect.