Mark all SFTP coroutine functions explicitly
This commit adds an asyncio.coroutine decorator to all functions in
SFTP which act as coroutines. Previously, some of these were plain
functions that returned a coroutine function (without yielding from
it). The new Python 3.5 "await" syntax is stricter about checking
things you await on and no longer allow this.