Move some processing from SSHServerConnection to create_server
This commit moves some processing previously done in the constructor
of SSHServerConnection to create_server, allowing any problems in
parsing the server host keys or authorized client keys to be detected
and reported immediately rather than triggering silent errors when a
new client connection is accepted.

This commit also allows the "host" argument on the listen function to be
defaulted (listening on all addresses) and removes some unnecessary
conditional checks found during unit testing. It also disables some code
which blocked password and keyboard-interactive auth when no cipher or
MAC was selected, since the current code doesn't allow these to be
disabled.

Finally, this commit disables branch coverage on a few conditional
code paths that currently will always be taken.