Complete adding unit tests for connection module
This commit adds the remaining tests necessary to get full coverage of
the connection module. With this check-in all modules except for sftp
are now at 100% code coverage under test.

This commit also refactors the server test module to move server
behavior specific to particular tests into the corresponding test
module, so both the client and server functionality needed for the
tests are kept together.