Update internal_error unit test to allow OS exceptions
The internal_error test causes the server to abruptly close the TCP
connection. On the client side, this can turn into either an AsyncSSH
DisconnectError or a lower-level OSError depending on where the
connection close is detected. This commit updates the unit test to allow
for either of these types of errors.