Make sure the right version of bcrypt is installed before using it
This commit changes the check for bcrypt to make sure that the version
installed is one that implements the KDF function before attempting to
use it for OpenSSH private key encryption/decryption. It also updates
the error raised when it is not available to indicate that bcrypt with
KDF support is needed.