Add support for Python 3.5 async context managers and update examples
This commit adds support for using Python 3.5 async context managers
(invoked with "async with") for SSH client connections, SFTP client
sessions, and SFTP file objects. It also adds support for using
"await" when creating these objects, and then later passing the result
into an "async with".

All of this has been done in a way which is backward compatible in both
Python 3.4 and Python 3.5 with the previous non-asynchronous context
managers (invoked with "with") and the older "yield from" syntax for
invoking coroutines.

This commit also updates the documentation and examples to all use the
new Python 3.5 syntax, and adds new tests to exercise the code which
supports this new syntax.