Add basic line editor functionality for SSH servers
To better support interactive SSH server applications, this commit adds
basic line editing and echoing capabilities for inbound SSH server
connections which request a pseudo-terminal. SSH clients opening such
connections generally deliver input a character at a time and expect the
remote server to provide input echo. AsyncSSH will now provide these
capabilities by default.

This new line editing feature can be disabled be setting the new
line_editor argument to False when starting the SSH server. It will also
automatically be disabled when the session encoding is set to None or
when the client does not request a pseudo-terminal.

When active, input data is delivered a line at a time to the server
application. The application can switch between line and character at a
time input using the new set_line_mode() function on the channel. Also,
echoing of input can be enabled or disabled using the new set_echo()
function.

Input editing works best on ANSI-compatible terminals. However, a more
basic form of line editing that doesn't wrap input across multiple lines
is also supported if the terminal type is not recognized.

Input history is also provided as part of this feature, with the ability
to tune the length of the history using the line_history argument when
starting the SSH server. Any input received when echo is disabled is
automatically excluded from the history.

This commit also fixes SSH server channels to not update terminal
information until after the session returns that it is ok to allocate a
pseudo-terminal. Previously, the requested terminal settings were saved
even when the SSH session refused the pseudo-terminal request.

This commit also changes the handling of newlines in output to server
channels with a pseudo-terminal. When this new feature is active,
newlines output to the channel are automatically converted to carriage
return line feeds so the application doesn't have to do this. This
makes the output more consistent when running with and without a
pseudo-terminal.

Documentation and examples have been updated to reflect these new
capabilities and a new example has been added to demonstrate the
set_line_mode() and set_echo() functions.

Unit testing is also included in this commit, with 100% code coverage
for the new functionality.