Add support for private key generation
This commit adds a generate_private_key() function which can be used
to create DSA, RSA, ECDSA, or Ed25519 private keys for use in SSH
public key authentication. Once generated, the resulting private keys
and the public keys corresponding to them can be exported in a variety
of formats for use in other tools.