Add support for generating and exporting SSH certificates
This commit adds support for generating new SSH certificates from an SSH
public key and a private signing key. It also adds the ability to export
SSH certificates in OpenSSH or RFC 4716 format, complementing the
previous support to import certificates in these formats.

This commit also adds unit tests for all of the new functionality and
cleans up a few minor issues found in the existing documentation.