Fix a bug in generation of ECDSA certificates
This commit changes the way certificate handler classes are registered
and looked up to correct a problem in generating certificates for ECDSA
keys.

This also makes the format_name argument optional in the
export_private_key(), export_public_key(), and export_certificate()
methods, defaulting it to 'openssh', and adds a declaration of the
return value of these functions in the documentation.