Fix an issue with ECDSA signature algorithms
The last commit to add support for alternate signature algorithms broke
the advertising of ECDSA public key algorithms in the list of supported
host key algorithms. This commit fixes this, and also fixes an issue
where the ordering of the ECDSA algorithms in the host key algorithm
list was not always consistent.