Add support for SSH extension negotiation
This commit adds supoprt for advertising SSH extensions as described in
https://tools.ietf.org/html/draft-ietf-curdle-ssh-ext-info-01. When
acting as a client, AsyncSSH will request a list of supported extensions
from the server, and when acting as a server it will advertise supported
extensions to clients which request this.

At the moment, the only supported extension is server-sig-algs, where
the server advertises a list of public key signature algorithms it
supports.

Since AsyncSSH doesn't currently understand any client extensions, it
does not request for the client to send such a list. However, all of the
framework necessary for doing so is in place and this can easily be
enabled when it is needed.