Change SFTP unit tests to avoid race condition on access time
The unit test which checked if a file was copied property and its access
time was preserved was re-opening the source and destination files to
compare their contents before checking their access times. However, the
act of opening the file changed their access time, causing the attribute
check to sometimes fail. This commit re-orders the check to test access
and modify times first and THEN check the file contents.