Rework alternate signature algorithm support
This commit reworks support for alternate signature algorithms. Instead
of passing the desired signature algorithm in as an additional argument
when loading individual keys, the list of allowed signature algorithms
is passed in when creating a new SSH client or server, similar to the
lists of key exchange, encryption, MAC, and compression algorithms. This
allows this list to be communicated using the newly supported ext_info
extension and reduces the number of separate authentication attempts
when using public key authentication, avoiding a possible issue where
the maximum allowed attempts is exceeded.