Complete the SSH agent client implementation
This commit completes the implementation of the SSH agent client code
in AsyncSSH. Now, in addition to being able to query keys stored in the
agent and offload signing operations to the agent, AsyncSSH supports
adding and removing keys from the agent (including keys with certificates
and constraints on key usage) and locking and unlocking access to the
agent. It also supports adding and removing smart card keys.

As part of adding this support, the APIs for loading local keys have
been enhanced to support encoding and decoding comment strings
associated with the keys, and passing these comment string to the agent
when keys are added. The key and certificate generation functions have
also been enhanced to support specifying a comment associated with the
newly created key/certificate.

Other accessor functions have also been aded to query what key algorithm
is associated with a key, certificate, or key pair, and to query whether
a key pair references a locally loaded key or an agent key.