Another fix for preserving access time in SFTP
In the previous commit to fix a race in preserving file access times
in SFTP, the code didn't take into account that the act of copying
the file changed the access time on the source system. So, to properly
preserve the access time at the destination, this commit sets the
access time on the destination file to that of the source file after
the copy is performed rather than its original access time.

This commit also updates the unit tests to intentionally set very old
values for the access and modify files on the original source file, to
make it more likely that a failure will be detected if the correct
access and modify times are not mirrored. Previously, this test would
only fail intermittently, if the current time in seconds happened to
change in the middle of the test.