Remove documentation for SSHKeyPair attributes
This commit removes the documentation for SSHKeyPair attributes. While
these instance variables will remain public and be used in AsyncSSH
outside of this class, they are not meant to be reflected from
application code. New methods are now available for applications which
want to access things like the algorithm or comment associated with a
key pair.