Skip additional SSH agent related tests when ssh-agent isn't available
This commit skips the initialization of an ssh-agent instance in the
ServerTestCase code and makes a new agent_available() method available
there which can be used to skip tests which require ssh-agent to be
present, allowing the tests to run correct when ssh-agent is not
available on the system. This was missed in a previous check-in.