Set file encoding explicitly in Unicode test
This commit sets the file encoding explicitly to UTF-8 before writing
Unicode data to a local file, rather than depending on this being the
system default. This avoids an error running the unit tests on systems
which default to ASCII or some other encoding that doesn't support the
characters being used in the test.