Add support for X11 forwarding in SSHServerConnection
This commit adds arguments to the AsyncSSH create_server() function to
enable X11 forwarding if clients request it. When this feature is
enabled and a client requests X11 forwarding, the new get_x11_display()
method can be called on the corresponding SSHServerChannel to get an X11
display value which can be passed to X11 applications running on the
server. Any connections opened to this display will be tunneled back
over the SSH connection to be displayed on the X11 display associated
with the client.