Avoid exception being thrown on Windows if SSH_AUTH_SOCK is set
The SSH agent code in AsyncSSH only expected SSH_AUTH_SOCK to be set in
the environment (or passed in explicitly) on UNIX systems. If this was
set on Windows, it would try to call asyncio.open_unix_connection() and
raise an exception. This commit catches this error, ignoring the failure
when UNIX domain socket support is not found.