Extend server-side SSH agent forwarding
This commit extends server-side SSH agent forwarding support to listen
on a UNIX domain socket so that external applications spawned by a
server session can tunnel requests back over SSH to the client's
ssh-agent. Previously, there was support for AsyncSSH itself to
connect back to the client's agent, and this access could be used
to authenticate multiple chained SSH connections, but there was no
way to export this access to other applications running on the server
machine.

This commit also adds additional documentation about SSH agent and
agent forwarding support.