Handle OverflowError in port forwarding
If a port number outside of the range 0-65535 is passed to the socket
bind function on Linux, it returns an OverflowError instead of an
OSError. AsyncSSH wasn't catching this case, and this led to it
leaking a socket object. This could be triggered by an SSH client
passing an invalid port number in a port forwarding request.

This commit now properly handles OverflowError, translating it to an
OSError before returning it so that calling code doesn't need to be
updated to handle both errors.