Add support for readuntil() method on SSHReader
This commit adds support for a readuntil() method on the SSHReader
class. It mimics the behavior of the readuntil() added to StreamReader
in Python 3.5.2, with the following exceptions:

    * Like other SSHReader/SSHWriter functions, I/O will be strings
      when an encoding is set and bytes otherwise. This means that
      the separator must be specified as a string or as bytes as
      well based on this same criteria.

    * SSHReader doesn't yet support the limit argument to restrict
      the maximum amount of data which can be returned. This support
      will be added in a future commit.

This commit also updates the documentation on some of the other
SSHReader methods to better describe what happens when signals are
received in a stream.