Add support for Pageant agent on Windows
This commit refactors the AsyncSSH agent client code to add support for
connecting to a Pageant agent when AsyncSSH is run on Windows. It
automatically detects the platform it is running on and chooses the
appropriate type of agent depending on the platform.

On Windows, it is not necessary to provide an agent path as an argument
or in the SSH_AUTH_SOCK environment variable to enable this support. All
that is needed is for Pageant to be running.