Use IPv4 loopback addresses to make TravisCI happy
By default, AsyncSSH unit tests try to open a server listening on local
loopback addresses. This generally defaults to IPv6 in Python,
attemping to listen on address ::1. However, this seems to cause
problems in some cases in TravisCI.

This commit passes "family=socket.AF_INET" to create_server() to force
it to select IPv4 addresses. Hopefully, this will eliminate the sporadic
TravisCI failures.