Add support for UMAC cryptographic hashes
This commit adds support in AsyncSSH for OpenSSH-compatible versions of
the UMAC cryptographic hashes defined in RFC 4418, based on the
specification at https://tools.ietf.org/html/draft-miller-secsh-umac-01.
However, the names advertised are suffixed with '@openssh.com' to allow
interoperability with OpenSSH (which doesn't advertise the unqualified
names proposed in the draft). Also, to match OpenSSH, only the 64 and 128
bits versions of UMAC are currently registered (matching OpenSSH), though
it would be very simple to add the 32 and 96 bit variants if a need arose.

The "etm" versions of these MAC algorithms are also supported, where
the MAC is calculated after encryption.

The UMAC support implemented here depends on the "nettle" library. These
new MAC algorithms will automatically become available when this library
is installed. Otherwise, only HMAC-based algorithms will be available.