Fix issue with loading nettle library on Windows
The ctypes.util.find_library() and ctypes.cdll.LoadLibrary() functions
behave differently on Windows and UNIX platforms when a library isn't
found, causing an unexpected TypeError exception. Also, find_library()
on Windows seems to require the full DLL name including the 'lib'
prefix and the version number suffix when they are present in the DLL
name, making find_library() not very useful there.

This commit updates the umac module to load the nettle library in a
platform-specific fashion to address these differences.