Add append functions for private/public keys and certificates
This commit adds append_private_key(), append_public_key(), and
append_certificate() methods on the correspond key and certificate
classes. These methods are similar to the existing "write" functions,
but append to the target file rather than overwriting it to allow a
list of keys/certificates to be written out.