Initial pass on making unit tests run on Windows
This commit is a first cut at getting the AsyncSSH unit tests to run on
Windows. It translates calls that relied on UNIX shell commands to
Python equivalents, and segregates tests that rely on UNIX-only Python
functionality so they can be conditionally skipped based on platform.

For now, the SFTP and X11 tests are disabled on Windows pending
additional analysis, but all but a handful of the remaining tests are
now running there.