Don't use ssh-keygen on Windows if it is installed
This commit disables the use of ssh-keygen in unit tests on Windows.
There seems to be a problem with it handling some key formats properly,
even though a similar version of ssh-keygen on UNIX systems works just
fine. With this change, the Windows unit tests will always fall back to
testing AsyncSSH against itself, rather than testing against ssh-keygen
if it is installed.