Re-enable X11 unit tests on Windows and allow forwarding without xauth
After looking more closely, the only problem getting the X11 code to run
on Windows turned out to be that os.rename() can't overwrite an existing
file there. To do a rename that replaces an existing file, os.replace()
must be used.

In addition, this commit allows X11 forwarding to be attempted even when
no xauth entry can be found in .Xauthority. In this case, a random auth
cookie is created and passed to the server, but connections will still
succeed if the server doesn't require authentication. This only applies
when a matching auth cookie can't be located for the server.

Unit tests have also been updated to no longer expect a failure in cases
where the auth token isn't found or .Xauthority is missing or empty.