Break readdir responses into chunks to avoid max message size limits
This commit changes the implementation of readdir() in the SFTP server
code to return 128 entries at a time rather than all of the entries in a
directory at once to avoid hitting limits in SFTP clients on the maximum
size of a response they will accept. Thanks go to Ben Leslie for
identifying the problem and suggesting this fix!