Expose overridable methods to format user & group names in SFTP server
This commit adds new format_user() and format_group() methods to the
SFTPServer class, allowing callers to provide a custom mapping from
uid/gid values to user/group names. The default remains the same as
before, using the pwd and grp modules to do the mapping on platforms
which have them, or returning the numeric values as a string
otherwise. Thanks go to Ben Leslie for this suggestion!