Fix issue updating environment in unit tests
This commit changes the way the process environment is updated so that
the the new environment is properly inherited by child processes. A
previous fix to avoid inheriting environment variables from the shell
used to launch the tests replaced os.environ entirely rather than
updating the values inside it, leading to problems with child processes
not picking up the new environment correctly.