Add support for GSSAPI key exchange and authentication
This commit adds support for GSSAPI-based key exchange methods and
gssapi-keyex and gssapi-with-mic authentication. This allows Kerberos
credentials to be used, as well as potentially other forms of
authentication which support GSSAPI. Both client and server side
are supported, as is support in the client to request delegated
credentials on the remote system after the authentication succeeds.

The "gssapi" Python module must be installed on UNIX systems for this
support to be available. Windows support is not available yet, but may
be added in the future.

In addition, this commit adds support for group 15 and group 17 key
exchange algorithms for both Diffie Hellman and GSS.