Add support for GSSAPI key exchange and authentication on Windows
This commit extends the recently-added GSSAPI support for UNIX to also
work on Windows, if the "sspi" module from pywin32 is installed.

Unlike UNIX, this support is limited for now to only supporting
Kerberos, as there doesn't seem to be a good way to determine the list
of supported mechanism OIDs. However, support for other security
packages could be added explicitly in the future by hardcoding other
package names and OIDs.

This commit is capable of supporting both client and server GSSAPI key
exchange and authentication. However, server support is only available
when running the application as a privileged user such as LocalSystem,
so that the host credentials are accessible.