Add optional progress reporting in SFTP client
This commit adds new block_size and progress_handler arguments to the
SFTP get, put, copy, mget, mput, and mcopy functions, allowing the
caller to control the block size used for file reads and writes and
to specify a callback function which can be used to track progress of
file transfers. The callback is called each time a block is copied
and it reports the source path, destination path, bytes copied so
far, and total bytes in the file being copied. If a request is made to
copy multiple files, the reporting is done consecutively on each file.