Add server-side class which supports I/O redirection
This commit adds an SSHServerProcess class which supports I/O
redirection of inbound connections to an SSH server. Instead of setting
the session_factory argument in the call to create_server(), you can set
the process_factory argument to have the server instantiate an
SSHServerProcess for each new inbound SSH session. This process class
supports the same I/O redirection options for stdin, stdout, and stderr
that were previously available through the create_process() function and
SSHClientProcess classes on the client side.

This commit also updates the examples to take advantage of this new
mechanism and adds a new example of server-side redirection using it.

Finally, this commit changes the preferred API from having separate
redirect_stdin(), redirect_stdout(), and redirect_stderr() methods to a
single redirect() method that can redirect one or more of these streams
in a single call. The old methods are still available for backward
compatibility, but the new more flexible redirect() method is preferred.