Fix SSH connection with known hosts object.
The connection was failing when the "known_hosts" parameter was a
SSHKnownHosts object previously imported from a string by calling
"import_known_hosts()", as documented in:

  http://asyncssh.readthedocs.io/en/latest/api.html#specifying-known-hosts

Python was throwing an exception since the SSHKnownHosts object was not
unpackable. Using the "self._known_hosts.match" method, which returns
the proper tuple, appears to fix the error.