Add SCP support
This commit adds support for an SCP client and server to AsyncSSH.

The SCP client can be used to get or put files, or to copy files
from one remote system to another. Wildcard glob matching is
supported, as are the same sort of error and progress handlers
supported by the SFTP client.

The SCP server support leverages the existing SFTPServer class to
allow for custom file handling on incoming SCP requests. All that
is required is to set "allow_scp" to True along with setting
"sftp_factory" and the SFTP server will also begin accepting SCP
requests.