Improve known_hosts support
This commit fixes AsyncSSH to properly accept a previously loaded
SSHKnownHosts object as the value of the known_hosts argument in the
create_connection() call. It also adds the ability to specify your
own matching function which takes the target host, address, and port
being connected to and returns the set of trusted host keys, trusted
CA keys, and revoked keys to use during host validation.