Expand set of values in CompletedProcess and ProcessError classes
This commit adds "env", "command", and "subsystem" properties to the
SSHClientProcess, SSHServerProcess, SSHCompletedProcess, and ProcessError
classes, allowing callers to retrieve information about what the process
was running. It also adds "stdout" and "stderr" properties to ProcessError,
allowing output to be examined when a process raises an error.

As part of this change, the methods get_environment(), get_command(),
and get_subsystem() previously available on SSHServerProcess should be
considered deprecated in favor of these new properties and may be removed
in a future release.

This commit also makes the get_environment(), get_command(), and
get_subsystem() methods on  SSHServerChannel also available on
SSHClientChannel.

Thanks go to iforapsy on GitHub for this suggestion!