Add support for trusted X.509 certificate "hash directories"
This commit adds support for an x509_trusted_cert_paths argument to
the create_connection() and create_server() functions. This argument
lets you specify a list of path names to directories which contain X.509
certificates which can be used for client or server authentication. To
facilitate efficient building of certificate trust chains, each
certificate in the directory is stored in a separate file, with a name
of the form <hash>.<number>, where <hash> is the OpenSSL hash value of
the certificate subject name and <number> is an integer counting up
from zero if multiple certificates in the directory have the same
subject hash.

For outbound client connections, this new argument defaults to the
directory .ssh/crt in the user's home directory. Any certificates
found in that directory can be used to build an X.509 trust chain
used to validate a server X.509 host certificate.

This argument is also available for validating X.509 client certificates
on inbound server connections, but no paths are added by default in this
case.