Replace local PBKDF2 implementation with one from PyCA
This commit removes a pure-Python PBKDF2 implementation and replaces it
with a call to a much faster version provided by PyCA.

In addition, this commit removes support for sha512-224 and sha512-256
truncated hash algorithms for private key encryption. As it turns out,
the previous implementation of these was incorrect, and so it only
worked if AsyncSSH was used to both encrypt and decrypt the keys.
Support for these may be added back in the future, but at the moment
neither PyCA nor OpenSSL seems to support these algorithms.