Improve support for file-like objects in process I/O redirection
This commit improves support for custom file-like objects passed in as
stdin, stdout, and stderr in the SSHClientProcess and SSHServerProcess
classes.

With this change, the fileno() call is allowed to raise OSError if the
file doesn't have a traditional file descriptor. Also, support is
provided for both text and binary file objects, based on whether or not
the object has an 'encoding' member.