Update MacOS config for Travis CI
This commit updates the Travis CI config to add a Python 3.6.1 MacOS
build. It also updates all MacOS builds to use Xcode 9.1 and updates the
older Python versions to 3.4.6 and 3.5.3.

This commit also updates the tox config to make the Python 3.7
nightly build be only for Linux, and not MacOS or Windows for now.