Add note about lack of OpenSSH support for send_signal()
This commit adds a note to the send_signal() method on SSHClientChannel
that the message it sends isn't supported by OpenSSH's SSH server. As a
result, this call and the terminate() and kill() calls end up being
ignored.

This commit also fixes a cosmetic issue in a few places related to RFC
references with section numbers. Previously, Sphinx did not include
the section anchor in the visible text, but now it does. So, I've
rewritten these references to provide the explicit text I want it to
use.