Update agent unit tests to work around issue in OpenSSH 7.6p1
This commit updates the "test_unknown_key" SSH agent unit test to use
a valid but non-loaded key to trigger the SSH agent to return an error,
rather than passing in an unparseable key. This works around a change
made in OpenSSH 7.6p1 which causes its SSH agent to exit with a fatal
error when receiving an invalid key rather than returning an error
and continuing to accept new requests.

As a result of the SSH agent exiting, the AsyncSSH unit tests sometimes
reported a "No such process" error when cleaning up if run on a system
with OpenSSH 7.6p1 installed. With this change, this should no longer
happen.

This change in OpenSSH has been reported and will likely be fixed in
the next release, but in the meantime this should prevent the AsyncSSH
unit tests from failing.