Update handling of Unicode comments in keys and certificates
This commit makes AsyncSSH less sensitive to encoding of non-ASCII
characters in comment fields of keys and certificates. The
get_comment() and set_comment() methods now take an optional
encoding paramter which specifies what encoding to use when converting
comment data between bytes and Unicode strings, and a new method has
been added called get_comment_bytes() to return comments as a raw
byte string, without performing any Unicode decoding.

This change should make AsyncSSH support keys and certificates with
any encoding of comment strings, rather than being restricted to UTF-8,
as Unicode decoding is no longer performed at the time keys are loaded.
The encoding is only needed when passing in a Unicode string as a
comment to set on a key/certificate or when requesting that comment
data be returned as a Unicode string. In those cases, the encoding will
default to UTF-8 for backward compatibilty, but other encodings can now
be specified.

As before, comment data can be passed in as a byte string to avoid the
need to conform to any specific encoding, and the get_comment_bytes()
method can now be used to retrieve comment data without specifying an
encoding.