Reduce time it takes to run unit tests
This commit reduces the time to run unit tests by decreasing the number
of rounds used when running bcrypt encryption of OpenSSH format private
keys. By setting the number of rounds to 1 rather than the default of
128, the total time to run the public key unit tests dropped by more
than a factor of 10, and the time to run the full set of unit tests
dropped by more than a factor of 3.

Newer versions of bcrypt (starting with 3.1.3) output a user warning
when a low number of rounds is used unless you include an
"ignore_few_rounds" argument. So, with this change, AsyncSSH now depends
on bcrypt 3.1.3 or later so that it can specify this argument from
within its unit test code without explicitly checking the version first.

This change also updates tox.ini to include explicit version numbers on
dependencies.