Update reference to asyncio.async to avoid issue in Python 3.7
This commit changes the access to asyncio.async to use getattr() to
avoid triggering a syntax error in Python 3.7. This function is only
used in Python versions prior to 3.4.4, but the reference triggers a
syntax error when the module is loaded in Python 3.7 since "async" is
now a true keyword. Referencing this function using getattr() allows
the code to compile in both older and newer versions of Python.