Add nnon-blocking collect_output() method to SSHClientProcess
This commit adds a new function named collect_output() to the
SSHClientProcess class. It returns unread output sent by the process to
stdout and stderr, similar to what is returned by communicate().
However, unlike communicate(), it does not send EOF to stdin and does
not wait for the process to exit before returning. It is intended to be
used instead of read() when you want to pick up pending output without
blocking.