Allow non-UTF8 comment data when generating or importing X.509 certificates
This commit allows non-UTF8 comment data to be passed into the AsyncSSH
functions which generate X.509 certificates, if the data is passed as an
already-encoded byte string. Ths change was made previously for OpenSSH
certificates, but it didn't quite work right in the X.509 case.

As with OpenSSH certificates, a caller must call get_comment() was an
appropriate "encoding" argument to properly decode such comments, or
call get_comment_bytes() to get the comment as a raw byte string.

This should also work on externally generated certificates which are
imported into AsyncSSH, if they contain an nsComment certificate
extension.