Fix an issue where some unprocessed packets were not being logged
This commit fixes an issue where logging was not being done for packets
which bypassed processing, such as the first kex message in cases where
it was skipped or channel messages sent to an invalid channel number.
Other log entries were produced in some of these cases, but the packet
triggering those was not actually being logged. This corrects that
problem, and adds a notation to such packets indicating why they weren't
processed.