Rework handling of unprocessed packets to make it more maintainable
This commit consolidates received packet processing and logging of
unprocessed packets in a single location, simplifying the design
and making it less likely that future changes will miss this
logging.

In the process, I was also able to remove some ugliness in the previous
design where certain packet types had to be manually excluded from being
logged inside of process_packet().