Additional work to simplify packet logging
This commit continues to simplify the packet logging code, such that all
logging is isolated to the SSHConnection send_packet() and recv_packet()
methods for regular SSH logs, and something similar is done in the
SFTPHandler send_packet() and recv_packets() methods, with a bit of
special-casing for the FXP_INIT and FXP_VERSION messages, due to their
lack of a packet ID.

This commit also eliminates the need to explicit pass the array of
handler names to use for logging into the send_packet() function.
Calling functions which are derived from the SSHPacketLogger class can
simply pass "self", and the log calls are actually done by calling back
into this handler so that the right handler names are automatically
picked up.

Finally, this commit now properly catches a decode error when a channel
message is sent without a channel number, treating this as another
special case where we skip processing but log the packet with a note
attached before raising a disconnect error.