Add support for pathlib paths in SFTP
This commit adds support for using pathlib objects in calls to SFTP
methods, in addition to Unicode or byte strings. This is mainly
intended to be used when specify local source and destination paths,
but these objects can be used to construct remote paths as long as
POSIX-style pathlib objects are used and an appropriate path encoding
is set to handle the conversion from Unicode to bytes.