Change server EXT_INFO message to only be sent after first key exchange
This commit limits the sending of the server MSG_EXT_INFO with to
only after the first key exchange's MSG_NEWKEYS. Previously,
MSG_EXT_INFO was being sent after every key exchange, and that
caused OpenSSH to complain about a protocol error.

The recently published RFC 8308 makes it clear that this message should
only be sent after the first key exchange, and this commit updates
AsyncSSH to comply with that.